fix errors, refactor, on step click, an active step opens in the constructor

+++ frontend/src/shared/components/constructorContent/ConstructorContent.tsx
@@ -27,7 +27,7 @@ export const ConstructorContent: React.FC<Props> = ({ type, itemId }) => {
   const dispatch = useAppDispatch()
 
   const onDeleteItem = () => {
-    dispatch(deleteItem({ itemId, activeBlockId }))
+    if (activeBlockId) dispatch(deleteItem({ itemId, activeBlockId }))
   }
 
   return (
